refactor: move API keys to nodes app
