+++ Meshflow/packets/views.py
@@ -2,8 +2,8 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from nodes.authentication import NodeAPIKeyAuthentication
 from .serializers import PacketIngestSerializer
-from constellations.authentication import NodeAPIKeyAuthentication
 
 
 class PacketIngestView(APIView):
